+++ merlin/common/tasks.py
@@ -767,9 +767,9 @@ def mark_run_as_complete(study_workspace: str) -> str:
         A string denoting that this run has completed.
     """
     merlin_db = MerlinDatabase()
-    db_run = merlin_db.get_run_by_workspace(study_workspace)
-    db_run.run_complete = True
-    db_run.save()
+    run_entity = merlin_db.get_run_by_workspace(study_workspace)
+    run_entity.run_complete = True
+    run_entity.save()
     return "Run Completed"