+++ CHANGELOG.md
@@ -13,8 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `db_scripts/` folder containing several new files all pertaining to database interaction
   - `db_commands`: an interface for user commands of `merlin database` to be processed
   - `merlin_db`: houses the `MerlinDatabase` class, used as the main point of contact for interactions with the database
-  - `db_study`: houses the `DatabaseStudy` class, used specifically for interacting with study entries in the database
-  - `db_run`: houses the `DatabaseRun` class, used specifically for interacting with run entries in the database
+  - `study_entity`: houses the `StudyEntity` class, used specifically for interacting with study entries in the database
+  - `run_entity`: houses the `RunEntity` class, used specifically for interacting with run entries in the database
 - Added `backends/` folder containing a new OOP way to interact with results backend databases
   - `results_backend`: houses an abstract class `ResultsBackend` that defines what every supported backend implement in Merlin
   - `redis_backend`: houses the `RedisBackend` class that defines specific interactions with the Redis database